Several cleanups - renamed vkContextBase -> vkRendererBase to accurately describe its functionality. - renamed vkDeferredShadingContext -> vkDeferredRenderer to follow new convention and purpose - created .cpps for vkCubemap and vkPanoramicTexture (WIP). - renamed vk::CameraController -> vk::Input - removed vkGlobal.cpp and move beginSingleTimeCommand() and endSingleTimeCommand() into vk::util (which is a lot more clear architecturally as to why it's there) - removed unused AssetManager::SyncIO() declaration.
... and other little things.

Will probably do some other cleanups along the way, but now focus is IBL.
