+++ src-vulkan/headers/vkDescriptorBuffer.h
@@ -1,11 +1,5 @@
 #pragma once
 
-extern PFN_vkGetDescriptorSetLayoutBindingOffsetEXT g_vkGetDescriptorSetLayoutBindingOffsetEXT;
-extern PFN_vkGetDescriptorSetLayoutSizeEXT g_vkGetDescriptorSetLayoutSizeEXT;
-extern PFN_vkGetDescriptorEXT g_vkGetDescriptorEXT;
-extern PFN_vkCmdBindDescriptorBuffersEXT g_vkCmdBindDescriptorBuffersEXT;
-extern PFN_vkCmdSetDescriptorBufferOffsetsEXT g_vkCmdSetDescriptorBufferOffsetsEXT;
-
 namespace vk
 {
 	//row = frame
