+++ src-vulkan/headers/vkDevice.h
@@ -1,4 +1,11 @@
 #pragma once
+#include "vkBuffer.h"
+
+extern PFN_vkGetDescriptorSetLayoutBindingOffsetEXT g_vkGetDescriptorSetLayoutBindingOffsetEXT;
+extern PFN_vkGetDescriptorSetLayoutSizeEXT g_vkGetDescriptorSetLayoutSizeEXT;
+extern PFN_vkGetDescriptorEXT g_vkGetDescriptorEXT;
+extern PFN_vkCmdBindDescriptorBuffersEXT g_vkCmdBindDescriptorBuffersEXT;
+extern PFN_vkCmdSetDescriptorBufferOffsetsEXT g_vkCmdSetDescriptorBufferOffsetsEXT;
 
 namespace vk 
 {
