+++ src/headers/Application.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Timer.h"
 #include "Physics.h"
-#include "vkContextBase.h"
+#include "vkRendererBase.h"
 #include "TextureManager.h"
 #include "DescriptorManager.h"
 
@@ -13,7 +13,7 @@ public:
 
 	PhysicsSystem& GetPhysics();
 	Timer& GetTimer();
-	vk::ContextBase* GetVulkanContext() const;
+	vk::RendererBase* GetVulkanRenderer() const;
 
 	void run();
 
@@ -28,7 +28,7 @@ private:
 private:
 	Timer mTime;
 	PhysicsSystem m_physics;
-	std::unique_ptr<vk::ContextBase> m_vulkanGraphicsContext;
+	std::unique_ptr<vk::RendererBase> m_vulkanGraphicsContext;
 	AssetManager m_assetManager;
 	TextureManager m_textureManager;
 	DescriptorManager m_descriptorManager;
